refactor: 메서드 분리 및 테스트 이동 - 연산자 테스트인 경우 StringCalculator 테스트가 아니므로 분리시켰다. - try-catch 블록에 필요하지 않은 범위도 처리하므로 분리시킴
