+++ src/test/java/calculator/StringCalculatorTest.java
@@ -1,10 +1,6 @@
  package calculator;
 
-import calculator.type.Add;
 import calculator.type.Divide;
-import calculator.type.Multiply;
-import calculator.type.Subtract;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -14,13 +10,7 @@ import static org.assertj.core.api.Assertions.*;
 
  public class StringCalculatorTest {
 
-    StringCalculator calculator;
-
-    @BeforeEach
-    @Test
-    void setUp() {
-        calculator = new StringCalculator();
-    }
+    StringCalculator calculator = new StringCalculator();;
 
     @ParameterizedTest
     @CsvSource(value = {"1=1",
@@ -32,34 +22,6 @@ import static org.assertj.core.api.Assertions.*;
         int result = calculator.calculate(input);
 
         assertThat(result).isEqualTo(expected);
-    }
-
-    @Test
-    void add() {
-        int result = new Add().operate(10, 20);
-
-        assertThat(result).isEqualTo(30);
-    }
-
-    @Test
-    void subtract() {
-        int result = new Subtract().operate(20, 10);
-
-        assertThat(result).isEqualTo(10);
-    }
-
-    @Test
-    void multiply() {
-        int result = new Multiply().operate(10, 20);
-
-        assertThat(result).isEqualTo(200);
-    }
-
-    @Test
-    void divide() {
-        int result = new Divide().operate(20, 5);
-
-        assertThat(result).isEqualTo(4);
     }
 
     @DisplayName("0으로 나누면 예외 발생")
